Add a multimeter continuity test substitute
After using a scrap of tin foil in sketch_01, I cut a square to match
the size of my artwork and taped it to the back. I also taped a jumper
wire to it but wasn't sure if I taped it well enough to make a
continuous connection. I don't have a multimeter, so I wrote this
sketch to test my work. I used a ground pin on the NodeMCU and D1. I
connected the wire taped to the tin foil to one, and attached a jumper
to the other to tap against the other corner of the tin foil.