Capacitive sensing with website monitor
By this point I had cut another square of tin foil and taped it to my
wall in what was a reasonable final position. I also taped a jumper
wire to this and needed to check if I had taped it well enough to be
continuous. I didn't have a Micro-USB cable long enough to connect it
to my PC and reach the wall. I could have shut down my PC and plugged
it in closer, but I didn't think of this. When I was working on this it
was 2:30am so I couldn't borrow a multimeter. But wait, sketch_02
doesn't require a serial monitor, just power to the NodeMCU that I
could have fulfilled with the nearest wall outlet or a USB batter pack.
I guess I didn't think of this because it was early, instead I decided
to code up a website interface and zoom in large enough that I could
read the numbers off my monitor from the wall. I spent way too long
on this. It felt like way more than half an hour of work.

I tried the httpUpdate example sketch but couldn't get it working or
figure out what it was supposed to do. I briefly looked for a
WiFi-to-Serial method. Tried a meta tag with refresh but wanted it to
update faster than every second. I ended up referencing
https://circuits4you.com/2016/12/16/esp8266-web-server-html/
for some help on serving a separate "file" (webpage in a '.h') and used
a XMLHttpRequest and setInterval. I learned a lot about working with
strings and char*, using itoa to convert numbers to characters, and
how fast I can pile up requests before the ESP8266 can't respond to any.

In the end the connection was good, I put the artwork on top of the tin
foil on the wall with some foam spacers, loaded sketch_02 back on and
had a working (very large) button where pressing the two sheets of tin
foil together to make contact lit up the built-in LED.