Google Home Notifier example updated
After getting a working button with the artwork on the all, the next
step was to trigger Spotify to play a song when that happens. I thought
I could send a command to Google Home from my ESP8266 like
"play ___ on everywhere" and browsing the googlehome GitHub topic
(https://github.com/topics/googlehome) brought me across this project
(https://github.com/horihiro/esp8266-google-home-notifier). This sketch
is basically just the example from it, but I had to update it to
address my Google Home by IP address because it wasn't picking it up
by hostname. This just plays a message, it doesn't actually trigger any
action or command. It is cool though and useful as a proof of concept,
to have Google say "now playing ___ on everywhere", if a tad confusing.