Start testing the ESP32's touch sensor
I borrowed JP's ESP32 since a week ago but haven't been able to flash
it until now when he came over and suggested holding the BOOT button.
Turns out I have to hold it for 3 seconds after I see the
"Uploading..." text.

This is a slight improvement of the ESP32 TouchRead example sketch. I
changed the polling rate from 1hz to 5hz for better responsiveness in
debugging. Now it also prints out "touched!" when the read value is
below the ambient level. It hovers at 26 or 27 in my case, with the
sheet of tin foil behind the artwork and plastic in the frame, and a
single 5' long 30 gauge wire connected to pin 4. It drops to 20 when I
touch it with a finger, or 12 when I put my whole hand on the plastic.
It drops to 25 when I touch the frame, and 24 or 23 even when I just
hover my hand above the plastic!