Keep track of total number of presses
I lowered the required touch count to 3 (3 consecutive measurement
below 26 will be considered a single touch) and started keeping track
of the total number of presses. I left it running overnight and there
wasn't a single additional press. The means that my limits are probably
good and won't randomly trigger and start playing music in the middle
of the night when I'm not using the art wall.