Try using PlatformIO for TouchInterrupt example
I wanted to start using PlatformIO for the folders in this project,
but it doesn't play nice with WSL.

I modified the example to only have one touch trigger, VS Code made it
super easy to rename the symbols. I also changed the pin to T0 because
that's what I currently have my frame tied to.