Remove debug print statements
I tried to see if both parts of the while condition really were necessary. If I only left client.connected() then the while loop would never exit. If I only left client.available() then Access Token would remain blank. So I put it back to the way it was and removed my debugging lines, content for now. It may be a placebo but it seems to help to wait a little while before accessing the esp32.local for the first time after formatting.