Working code to get currently-playing album art
Note a 204 means nothing is playing. I'm not sure if the || to && change is needed in the update function too, but it works with it. I'm leaving in the code I used to test starting from scratch, SPIFFS.remove didn't seem to be enough to make it not say "Using refresh token found on the FS".