Working Spotify interactivity
Rather than play/pausing every 10 seconds, attach it to the touch sensor code from sketch_07. I'm using the interrupt as opposed to the read from sketch_06 because the update and other http calls take time that the loop won't be running and might miss a touch. I got double taps (pause then immediately play) when I had touchDetected = false; as the fire line in the if statement, moving it to the end seems to resolve the issue.