Remove device_id from playSong and add debugging
I was experimenting with calls in the Spotify developer documentation
and Postman. It turns out the Google Homes do not show up in the list
of devices unless one is currently playing music. If its device id
is provided anyway, the call will fail. This means my best course of
action is likely using the Google Assistant API to ask it to play a song
"on everywhere" direction, that or some hacky solution with Spotify.

I also learned that there's an API to play a song immediately so I don't
have to queue up a song and skip forward, not that that matters much now
that I'll be using the Google Assistant API instead of Spotify's.