Add test animation on ESP32 with display
I used youtube-dl to download the video (https://youtu.be/Qj6a10yZUaY),
After Effects to scale it to 135x240 and export to a jpeg sequence,
dupeGuru to find and delete duplicates,
but scrapped that idea and just uploaded all but the last 33 frames
that wouldn't fit.

I used https://youtu.be/UE1mtlsxfKM as the tutorial,
https://github.com/me-no-dev/arduino-esp32fs-plugin/releases/ to upload
the images, and the SPIFFS_Jpeg example from the JPEGDecoder library
as my base.

At ~50ms render time we get 20fps.